D: integer coordinates in the model - no floats below the graphics boundary
Pos is now { x: i32, y: i32 }. The model carries no floats at all: footprints,
overlap, and placement are exact integer geometry, and a coordinate becomes a
float only on the Bevy side, which converts at the boundary because the graphics
card wants floats. This makes the rules/presentation separation literal - the
model is pure integer 2-space; the renderer owns the pixels.

- Pos -> i32 (now also Eq/Hash); dist_sq -> i64 (a coordinate span squared
  overflows i32) and becomes a total order, simplifying the separate() sort.
- geometry: overlap area and distance widened to i64; the epsilon float
  tolerances become exact integer comparisons (no more 0.01 slop).
- layout: all card/chip/cascade constants and footprints are i32.
- physical: set_pile_pos/set_card_pos/place_pile/set_bounds/set_pile_size and
  arrange_row/structured_positions take i32; default_bounds is 1_000_000.
- renderer + fixtures: convert f32->i32 when feeding the model (dropped pixels
  round at the boundary) and i32->f32 when reading a position for Val::Px;
  animation still lerps in f32. GAP/OVERLAY_BAND/CARD_W/CARD_H are i32 (they
  feed the model); the pixel-drawing aliases (SMALL_W, ...) stay f32.

+++ crates/cardtable-model/src/binding.rs
@@ -34,7 +34,7 @@ fn add_zone(tree: &mut Board, parent: PileId, zone: &ZoneView, index: usize) -> 
     let pile = tree
         .add_pile(parent, zone.label.clone())
         .expect("parent pile exists");
-    tree.set_pile_pos(pile, 24.0 + index as f32 * 180.0, 24.0)
+    tree.set_pile_pos(pile, 24 + index as i32 * 180, 24)
         .expect("just-created pile exists");
     // Honour the card-table arrangement (grid / rows / free). `List` is the model default that
     // `add_pile` already sets, so only override otherwise; preserve the pile's `editable` flag (the seam
